requirements.txt
python-telegram-bot==22.1
fastapi
uvicorn
gspread
oauth2client
pandas
